proxy-function/requirements.txt
functions-framework>=3.0.0
google-cloud-workflows>=1.0.0
google-cloud-vision>=3.0.0
google-cloud-documentai>=2.0.0
google-cloud-videointelligence>=2.0.0
google-cloud-aiplatform>=1.0.0
Pillow>=10.0.0
